Better handling of broken workdir seperation
In case S is set to WORKDIR the finding of files does include
folders like recipe-sysroot and recipe-sysroot-native.
This leads to false positives as these files are not to be checked
in the context of the current recipe.
Also this could lead  to issue when exporting sources when
SCA_EXPORT_FINDING_SRC is set. E.g. with PermissionError
or similar.
To solve this issue filter out the files in first place by using
the new variable SCA_LOCAL_FILE_FILTER in case SCA_SOURCES_DIR
is the same as WORKDIR and do exceptions handling
on coying files.

Closes #236
Closes #237