Be more picky on mode selection
In before the MODE of SCA (image or recipe) was determined by
probing for var IMAGE_INSTALL being set to a none empty value.

As some recipe seem to set this variable although they are NOT
image recipe, we need to be more picky.

Now IMAGE_INSTALL and IMAGE_TYPES need to be set to a non empty
value before the recipe is consired an image, everything else in
run in recipe mode.