Fix issue on native recipes
If being invoked on to native or cross recipes helper.bbclass was
setting a DEPENDS on python3-magic-native, which requires
python3-native. So if is code was run on python3-native it self we had
a dependency loop.
Fixed this be using blacklist function on this particular DEPENDS as well.

Rename sca-blackllist to sca-blacklist for obvious reasons