Using prefuncs and postfuns to inject sca-tasks
Remove the usage of addtask, which causes the hashes
of internal bitbake hashing to change, so recipes
which inherit any sca-task are constantly rebuild.

Use pre+postfuncs instead which don't seem to have
that side effect - another benefit is that we are
no more able to control when a task is actually executed.

Closes #372