Fix xmllint suppress
The xmllint tool generates error messages on the form 'validity error', which are transformed into datamodel id's of 'xmllint.xmllint.validity error'.

The embedded space in the id strings makes it impossible to suppress error messages with SCA_XMLLINT_EXTRA_SUPPRESS.

Change the regex to not capture the trailing ' error' from the xmllint output, which will make it possible to suppress xmllint warnings.