Fix force run feature
In before the SCA_FORCE_RUN feature worked
with the help of conditionally tagging the
nostamp flag at all sca-tasks.
It turned out that the value of the flag isn't of
any important. The presence of the flag is
enough to tag the task, which let to the effect
that all sca-tasks were executed over and over
again although the wasn't need to do so, as
none of the input vars had changed.

The new system is working with a dummy task,
which is tagged statically with nostamp.
This nostamp class is inserted as a dependency
to all sca-tasks based on the SCA_FORCE_RUN
variable.

Currently this works for all modules expect
latest cve-check, which will be dealt with in
another commit

Closes #372