Fix cve-check issue
The integration of cve-check class requires for whatever
reasons a global inherit - this is not how things work with
this layer.
In cve-update-db-native.bb a hardcoded check is included,
which probes for vars from cve-check class being included.

So we work around this whole issue by setting the required
vars in an bbappend for cve-update-db-native

Closes #477