Override emit_pkgdata at python
This addresses the issue with packaging
not existing pyc-files in python3.
As the underlying issue is clearly to be
found in poky itself, we apply only a
hotfix here, which just ignores any
NOT found files while emit_pkgdata.
This should make building work again,
without interfering too much with other
components.
Therefore the hotfix is only applied to python3

Relates to #649