Add spare layer/dir support
This replaces the anonymous python func for
primary invoking the SCA by a bitbake handler function.
This has to be done, as the order of anon python function
is not predictable esp. when bbappends are included,
which otherwise led to missing append files - but they
are important.
Also it should be more bitbakeish to do so.

Closes #899