Remove lock and sock files
Sometimes it could be that bitbake is
unable to restart, which is attempted when
a lock and/or sock file is found in the build-dir.

Later on this srews the whole workflow but,
as unforseen events such as out of disk space
could happen.

To prevent that remove both of the files
at every possible place throughout the workflow.

With the files gone a "clean" bitbake instance
is used, which is pretty much that what we want