crossemu: handle pseudo
pseudo is a set up as a singe instance server per workspace, creating a
sqlite db of all operations performed under the virtual fakeroot (sigh).

Somehow the server isn't shutdown after a task is done, but kept in
the workspace running.
If we now create plenty of side rootfs, this will inevitably to
collisions, as the README of pseudo states the used db is prone
to miss out on changes when put under pressure (which is what we
exactly do).
Upstream is using a shutdown flag (-S) to make the server quit
between tasks - unfortunately this doesn't work reliable for
python based task, as the fakeroot in injected on a bitbake
worker level and not on a task/func level as with shell functions.

To workaround that fact, we need to reintroduce a file lock, but this
time not for the whole task but only for the do_rootfs part.
Before and after the rootfs creation pseudo server is forefully killed
and all leftover files (db and stuff) are deleted.
server will autospawn again starting with a blank db, which is exactly
what we want.

Related to #4680

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>