remove python3-nose
nose test framework is highly deprecated, having been unmaintained for a
long time. There is just one dependency on nose in this layer,
python3-flake8-logging-format, which turns out to be a spurious
dependency that can be fixed by a simple patch.

python3-flake8-logging-format incorrectly listed nose as a
standard requirement, rather than listing it with the test dependencies,
but nose is clearly a test requirement *only* used when running the
unit tests. No functionality of the package itself depends on nose,
so moving it to test_require is the right thing to do for now
(i.e. until someone updates the unit tests to run under a currently
supported test framework).

Signed-off-by: S. Lockwood-Childs <sjl@vctlabs.com>
Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>