go: explicitly set branch for remotes without tags
in case we are using UPSTREAM_CHECK_COMMITS with a git
fetcher we can't use nobranch at the same time,
as the bitbake code simply wouldn't know what ancestors to
this particular commit exist.
So in this case we have to explicitly set a remote branch to
pass the upgrade-check-status from bitbake.

Closes #9474

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>