Uncontionally build everything in CI
As the choices of SCA modules influence the sstate-cache,
a changing set influences the cache reusage, rending the
pre-made cache pretty much useless.
To avoid this put every module into a standard set and
run it unconditionally.
This also adds module that were in before not part of the
default set of SCA modules.

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>