classes: set proper vardeps for tasks
for all variables that are computed or using clean_split function,
as they can't be seen by the bitbake parser.
Like this e.g. a change to SCA_SCOPE_FILTER in local or
distro.conf will rerun all the tasks that inherit that variable.

Along the way fix the following errors

- kconfighard: allow for all kernel recipes not just linux-yocto
- perlcritic: fix shebang pattern
- bandit: fix SCA_BANDIT_EXTRA_FATAL usage
- it: exclude BB_NUMBER_THREADS from task hash
- multimetric: weak define SCA_MULTIMETRIC_COMPILER_MODULES
- shellcheck: weak define SCA_SHELLCHECK_EXTRA_* vars

Closes #11833

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>